workflow fixes; update makefile

+++ tools/remove_dependencies.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-# Copyright 2019 The Vitess Authors.
+# Copyright 2023 The Vitess Authors.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -16,9 +16,7 @@
 
 # Remove tools and dependencies installed by "make tools"
 
-get_arch() {
-  uname -m
-}
+set -euo pipefail
 
 function fail() {
 	echo "ERROR: ${1}"
@@ -27,26 +25,29 @@ function fail() {
 
 BUILD_JAVA=${BUILD_JAVA:-1}
 BUILD_CONSUL=${BUILD_CONSUL:-1}
-VTROOT="$PWD/../"
-
-[[ "$(dirname "$0")" = "." ]] || fail "remove_dependencies.sh must be run from its current directory"
+uname=$(uname)
+get_arch=$(uname -m)
 
 uninstall_protoc() {
     echo "Removing protoc..."
     local dist="$1"
 
-    if [ -f $dist ]; then
+    if [ -f "$dist/bin/protoc" ]; then
         unlink "$dist/bin/protoc"
         rm "$VTROOT/bin/protoc"
     fi
-    rm -rf $dist
+    if [[ "$(echo $dist | awk -F/ '{print $NF}')" == "vt-protoc-$PROTOC_VER" ]]; then
+        rm -rf $dist
+    fi
 }
 
 uninstall_zookeeper() {
     echo "Removing zookeeper..."
     local dist="$1"
 
-    rm -rf $dist
+    if [[ "$(echo $dist | awk -F/ '{print $NF}')" == "vt-zookeeper-$ZK_VER" ]]; then
+        rm -rf $dist
+    fi
 }
 
 uninstall_etcd() {
@@ -54,17 +55,17 @@ uninstall_etcd() {
     local version="$1"
     local dist="$2"
 
-    case $(uname) in
+    case uname in
         Linux)  local platform=linux; local ext=tar.gz;;
         Darwin) local platform=darwin; local ext=zip;;
-        *)   echo "ERROR: unsupported platform for etcd"; exit 1;;
+        *)   echo "Etcd not installed. Ignoring.."; return;;
     esac
 
-    case $(get_arch) in
+    case get_arch in
         aarch64)  local target=arm64;;
         x86_64)  local target=amd64;;
         arm64)  local target=arm64;;
-        *)   echo "ERROR: unsupported architecture for etcd"; exit 1;;
+        *)   echo "Etcd not installed. Ignoring.."; return;;
     esac
 
     if [ -f "$dist/etcd-${version}-${platform}-${target}/etcd" ]; then
@@ -75,7 +76,10 @@ uninstall_etcd() {
         unlink "$dist/etcd-${version}-${platform}-${target}/etcdctl"
         rm "$VTROOT/bin/etcdctl"
     fi
-    rm -rf $dist
+
+    if [[ "$(echo $dist | awk -F/ '{print $NF}')" == "etcd" ]]; then
+        rm -rf $dist
+    fi
 }
 
 uninstall_consul() {
@@ -86,24 +90,26 @@ uninstall_consul() {
         unlink "$dist/consul"
         rm "$VTROOT/bin/consul"
     fi
-    rm -rf $dist
+    if [[ "$(echo $dist | awk -F/ '{print $NF}')" == "consul" ]]; then
+        rm -rf $dist
+    fi
 }
 
 uninstall_toxiproxy() {
     echo "Removing toxiproxy..."
     local dist="$1"
 
-    case $(uname) in
+    case uname in
         Linux)  local platform=linux;;
         Darwin) local platform=darwin;;
-        *)   echo "WARNING: unsupported platform. Some tests that rely on toxiproxy will not function."; return;;
+        *)   echo "Toxiproxy not installed. Ignoring.."; return;;
     esac
 
-    case $(get_arch) in
+    case get_arch in
         aarch64)  local target=arm64;;
         x86_64)  local target=amd64;;
         arm64)  local target=arm64;;
-        *)   echo "WARNING: unsupported architecture. Some tests that rely on toxiproxy will not function."; return;;
+        *)   echo "Toxiproxy not installed. Ignoring.."; return;;
     esac
 
     file="toxiproxy-server-${platform}-${target}"
@@ -112,25 +118,25 @@ uninstall_toxiproxy() {
         unlink "$dist/$file"
         rm "$VTROOT/bin/toxiproxy-server"
     fi
-    rm -rf $dist
+    if [[ "$(echo $dist | awk -F/ '{print $NF}')" == "toxiproxy" ]]; then
+        rm -rf $dist
+    fi
 }
 
 uninstall_all() {
     echo "## local system details..."
-    echo "## platform: $(uname) target:$(get_arch) OS: $os"
+    echo "## platform: $uname target:$get_arch OS: $OSTYPE"
 
     # protoc
-    protoc_ver=21.3
-    uninstall_protoc "$VTROOT/dist/vt-protoc-$protoc_ver"
+    uninstall_protoc "$VTROOT/dist/vt-protoc-$PROTOC_VER"
 
     # zk
-    zk_ver=${ZK_VERSION:-3.8.0}
     if [ "$BUILD_JAVA" == 1 ] ; then
-        uninstall_zookeeper "$VTROOT/dist/vt-zookeeper-$zk_ver"
+        uninstall_zookeeper "$VTROOT/dist/vt-zookeeper-$ZK_VER"
     fi
 
     # etcd
-    uninstall_etcd "v3.5.6" "$VTROOT/dist/etcd"
+    uninstall_etcd $ETCD_VER "$VTROOT/dist/etcd"
 
     # consul
     if [ "$BUILD_CONSUL" == 1 ] ; then
